+++ src/app/content/highlights/components/DisplayNote.tsx
@@ -7,8 +7,9 @@ import Times from '../../../components/Times';
 import { textStyle } from '../../../components/Typography/base';
 import theme from '../../../theme';
 import { mergeRefs } from '../../../utils';
+import { highlightStyles } from '../../constants';
 import { focusHighlight } from '../actions';
-import { cardPadding, cardWidth, highlightStyles } from '../constants';
+import { cardPadding, cardWidth } from '../constants';
 import Confirmation from './Confirmation';
 import MenuToggle, { MenuIcon } from './MenuToggle';
 import TruncatedText from './TruncatedText';
