+++ src/app/content/highlights/components/EditCard.spec.tsx
@@ -12,8 +12,8 @@ import { formatUser } from '../../../auth/utils';
 import * as Services from '../../../context/Services';
 import MessageProvider from '../../../MessageProvider';
 import { assertDocument } from '../../../utils';
+import { highlightStyles } from '../../constants';
 import { updateHighlight } from '../actions';
-import { highlightStyles } from '../constants';
 import ColorPicker from './ColorPicker';
 import EditCard, { EditCardProps } from './EditCard';
 import Note from './Note';
