+++ src/app/content/highlights/components/ShowMyHighlights.tsx
@@ -2,12 +2,12 @@ import { HTMLElement } from '@openstax/types/lib.dom';
 import flow from 'lodash/fp/flow';
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
+import GoToTopButton from '../../../components/GoToTopButton';
 import withServices from '../../../context/Services';
 import { isHtmlElement } from '../../../guards';
 import { AppServices, AppState, Dispatch } from '../../../types';
-import GoToTopButton from '../../components/GoToTopButton';
+import { loadMoreDistanceFromBottom } from '../../constants';
 import { loadMoreSummaryHighlights } from '../actions';
-import { loadMoreDistanceFromBottom } from '../constants';
 import * as select from '../selectors';
 import Highlights from './Highlights';
 import * as Styled from './ShowMyHighlightsStyles';
