+++ src/app/content/highlights/components/SummaryPopup/HighlightListElement.tsx
@@ -5,9 +5,9 @@ import styled, { css } from 'styled-components/macro';
 import { useAnalyticsEvent } from '../../../../../helpers/analytics';
 import { bodyCopyRegularStyle } from '../../../../components/Typography';
 import theme from '../../../../theme';
+import { highlightStyles } from '../../../constants';
 import { popupBodyPadding } from '../../../styles/PopupStyles';
 import { deleteHighlight, updateHighlight } from '../../actions';
-import { highlightStyles } from '../../constants';
 import ContextMenu from './ContextMenu';
 import HighlightAnnotation from './HighlightAnnotation';
 import HighlightDeleteWrapper from './HighlightDeleteWrapper';
