+++ src/app/content/highlights/hooks/initializeMyHighlightsSummary.ts
@@ -4,9 +4,9 @@ import pickBy from 'lodash/fp/pickBy';
 import { isDefined } from '../../../guards';
 import { ActionHookBody } from '../../../types';
 import { actionHook, assertDefined } from '../../../utils';
+import { summaryPageSize } from '../../constants';
 import * as selectContent from '../../selectors';
 import { initializeMyHighlightsSummary, receiveHighlightsTotalCounts, receiveSummaryHighlights } from '../actions';
-import { summaryPageSize } from '../constants';
 import * as select from '../selectors';
 import { CountsPerSource } from '../types';
 import { loadMore } from './loadMore';
