+++ src/app/content/highlights/hooks/locationChange.ts
@@ -3,10 +3,10 @@ import { getType } from 'typesafe-actions';
 import { receivePageFocus } from '../../../actions';
 import { user } from '../../../auth/selectors';
 import { AnyAction, AppServices, MiddlewareAPI } from '../../../types';
+import { maxHighlightsApiPageSize } from '../../constants';
 import { bookAndPage } from '../../selectors';
 import { Book } from '../../types';
 import { receiveHighlights } from '../actions';
-import { maxHighlightsApiPageSize } from '../constants';
 import * as select from '../selectors';
 import { HighlightData, SummaryHighlightsPagination } from '../types';
 import { extractDataFromHighlightClientResponse, incrementPage } from './utils';
