+++ src/app/content/highlights/hooks/printHighlights.spec.ts
@@ -8,6 +8,7 @@ import { resetModules } from '../../../../test/utils';
 import { MiddlewareAPI, Store } from '../../../types';
 import { assertWindow } from '../../../utils';
 import { receiveBook, receivePage } from '../../actions';
+import { maxHighlightsApiPageSize } from '../../constants';
 import { formatBookData } from '../../utils';
 import {
   closeMyHighlights,
@@ -18,7 +19,6 @@ import {
   setSummaryFilters,
   toggleSummaryHighlightsLoading,
 } from '../actions';
-import { maxHighlightsApiPageSize } from '../constants';
 import { HighlightData, SummaryHighlights } from '../types';
 
 const book = formatBookData(archiveBook, mockCmsBook);
