+++ src/app/content/highlights/selectors.spec.ts
@@ -2,11 +2,6 @@ import { treeWithoutUnits } from '../../../test/trees';
 import { book } from '../selectors';
 import * as select from './selectors';
 
-jest.mock('./constants', () => ({
-  enabledForBooks: ['enabledbook'],
-  highlightStyles: (jest as any).requireActual('./constants').highlightStyles,
-}));
-
 const mockBook = book as any as jest.SpyInstance;
 
 jest.mock('../selectors', () => ({
