+++ src/app/content/studyGuides/components/ShowStudyGuides.spec.tsx
@@ -5,11 +5,11 @@ import renderer from 'react-test-renderer';
 import createTestServices from '../../../../test/createTestServices';
 import createTestStore from '../../../../test/createTestStore';
 import { renderToDom } from '../../../../test/reactutils';
+import GoToTopButton from '../../../components/GoToTopButton';
 import * as Services from '../../../context/Services';
 import MessageProvider from '../../../MessageProvider';
 import { Store } from '../../../types';
 import { assertWindow } from '../../../utils';
-import GoToTopButton from '../../components/GoToTopButton';
 import { loadMoreStudyGuides } from '../actions';
 import * as selectors from '../selectors';
 import ShowStudyGuides from './ShowStudyGuides';
