+++ src/app/content/studyGuides/components/ShowStudyGuides.tsx
@@ -2,10 +2,10 @@ import { HTMLElement } from '@openstax/types/lib.dom';
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import styled, { css } from 'styled-components/macro';
+import GoToTopButton from '../../../components/GoToTopButton';
 import theme from '../../../theme';
 import { assertNotNull } from '../../../utils';
-import GoToTopButton from '../../components/GoToTopButton';
-import { loadMoreDistanceFromBottom } from '../../highlights/constants';
+import { loadMoreDistanceFromBottom } from '../../constants';
 import { PopupBody } from '../../styles/PopupStyles';
 import { loadMoreStudyGuides } from '../actions';
 import * as select from '../selectors';
